requirements.txt
fastapi
uvicorn
pydantic>=2.0.0
pymongo
boto3
paho-mqtt
requests